Add school year to sidebar
